app/requirements.txt
streamlit
psycopg2-binary
sqlalchemy
sqlalchemy-redshift